Make ntpcheck work on freebsd and darwin
