+++ .circleci/config.json
@@ -2,7 +2,7 @@
     "gopkg": "github.com/facebookincubator/ntp",
     "licenses": [
         [
-            "^/\\*",
+            "^(// \\+build.+\n\n)?/\\*",
             "Copyright \\(c\\) Facebook, Inc\\. and its affiliates\\.",
             "",
             "Licensed under the Apache License, Version 2\\.0 \\(the \"License\"\\);",
